LayoutTests/editing/spelling/spelling-multiword-selection-expected.txt

Spelling should be disabled when user selects multiple words that are not a single misspelling. To test manually, type 'wellcome home.', select 'wellcome home', and context-click on the selection. The test suceeds when the context menu has no spell check suggestions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome home"
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is "wellcome home"
PASS contextMenuElements[contextMenuElements.length - 1] is "<separator>"
PASS successfullyParsed is true

TEST COMPLETE
wellcome home.
